agents/echo/requirements.txt
python-docx>=1.1.0
portal-sdk
